Refactor escaping in LBT
