+++ lib/lbt/bundle/AutoSplitter.js
@@ -3,10 +3,9 @@
 const uglify = require("uglify-es");
 const {pd} = require("pretty-data");
 
-const nonAsciiEscaper = require("../../processors/nonAsciiEscaper");
-
 const ModuleName = require("../utils/ModuleName");
 const {SectionType} = require("./BundleDefinition");
+const escapePropertiesFiles = require("../utils/escapePropertiesFiles");
 const log = require("@ui5/logger").getLogger("lbt:bundle:AutoSplitter");
 
 const copyrightCommentsPattern = /copyright|\(c\)(?:[0-9]+|\s+[0-9A-za-z])|released under|license|\u00a9/i;
@@ -229,26 +228,11 @@ class AutoSplitter {
 				return out.toString().length();
 				*/
 
-				// input encoding is read from project configuration, default is "ISO-8859-1"
-				// since AutoSplitter is also used when splitting non-project resources (e.g. dependencies)
+				// Since AutoSplitter is also used when splitting non-project resources (e.g. dependencies)
 				// *.properties files should be escaped if encoding option is specified
-				const propertiesFileEncoding = resource.getProject()
-					&& resource.getProject().resources
-					&& resource.getProject().resources.configuration
-					&& resource.getProject().resources.configuration.propertiesFileEncoding;
-				if (propertiesFileEncoding) {
-					const encoding = nonAsciiEscaper.getEncodingFromAlias(propertiesFileEncoding);
-					await nonAsciiEscaper({
-						resources: [resource.resource],
-						options: {
-							encoding
-						}
-					});
-				}
+				const fileContent = await escapePropertiesFiles(resource);
 
-				const fileContent = await resource.buffer();
-
-				return fileContent.toString().length;
+				return fileContent.length;
 			} else if ( this.optimizeXMLViews && /\.view.xml$/.test(module) ) {
 				// needs to be activated when it gets activated in JSMergedModuleBuilderExt
 				let fileContent = await resource.buffer();
