+++ lib/lbt/bundle/Builder.js
@@ -9,11 +9,10 @@ const escodegen = require("escodegen");
 const {Syntax} = esprima;
 // const MOZ_SourceMap = require("source-map");
 
-const nonAsciiEscaper = require("../../processors/nonAsciiEscaper");
-
 const {isMethodCall} = require("../utils/ASTUtils");
 const ModuleName = require("../utils/ModuleName");
 const UI5ClientConstants = require("../UI5ClientConstants");
+const escapePropertiesFiles = require("../utils/escapePropertiesFiles");
 
 const BundleResolver = require("./Resolver");
 const BundleSplitter = require("./AutoSplitter");
@@ -444,26 +443,11 @@ class BundleBuilder {
 			}
 			outW.write( makeStringLiteral( fileContent ) );
 		} else if ( /\.properties$/.test(module) ) {
-			// same as for other text files, input encoding is read from project configuration, default is ISO-8859-1
-			// since Builder is also used when building non-project resources (e.g. dependencies)
+			// Since the Builder is also used when building non-project resources (e.g. dependencies)
 			// *.properties files should be escaped if encoding option is specified
-			const propertiesFileEncoding = resource.getProject()
-				&& resource.getProject().resources
-				&& resource.getProject().resources.configuration
-				&& resource.getProject().resources.configuration.propertiesFileEncoding;
-			if (propertiesFileEncoding) {
-				const encoding = nonAsciiEscaper.getEncodingFromAlias(propertiesFileEncoding);
-				await nonAsciiEscaper({
-					resources: [resource.resource],
-					options: {
-						encoding
-					}
-				});
-			}
+			const fileContent = await escapePropertiesFiles(resource);
 
-			const fileContent = await resource.buffer();
-
-			outW.write( makeStringLiteral( fileContent.toString() ) );
+			outW.write( makeStringLiteral( fileContent ) );
 		} else {
 			log.error("don't know how to embed module " + module); // TODO throw?
 		}
